iOS/Android Wrappers (#13)
The wrapper code has been reviewed by at least 3 other working group members (Quantcast, Conversant, and Didomi IIRC), so as long as there are not conflicts in references, we should be fine to merge. If we do notice any errors because we’re pushing to merge today, we will connect next week to resolve. I will also be reviewing license documentation with a closer eye next week (thank you for your contribution to Tech Lab and for using the licenses already existing within the codebase).

* branch synced with master + in-App code

* No Smaato copyright. CC Licenses added