1. Change addEventListener/removeEventListener interface
window.postmessage() cannot serialize a function object, so the
original spec for removeEventListener does not work cross-frame.
To handle this, let CMP maintain a internal map of registered
listeners and their ID.

- TCData layout: add `listenerId`
- removeEventListener `parameter` param: `callback` -> `listenerId`

2. Clarify addEventListener callback invocation time

Emphasize that the callback should be invoked with the current TC
String upon registration, so that it is possible to recommend
"calling scripts register a listener function via addEventListener
instead of getTCData" as suggested by the getTCData section.

3. Remove SafeFrame proxy communications.

Remove descriptions on safeFrame proxy since it is incompatible
with caller sync expectations.