Handling conflict between v1.1 strings and v2 strings (#212)
* Handling conflict between v1.1 strings and v2 strings

During a Framework Signal WG call deprecation of TCF v1.1 strings was discussed and it was agreed that help manage this process and maintain the integrity of the TC String (the WG mission) that CMPs should be asked to remove TCF v1.1 strings if a TCF v2.0 string was present. This ensures that vendors are not conflicted by presentation of two strings. Policy have have confirmed that this does not require a policy change but we feel that this does require a Tech Spec change. Patrick is off this week so I am picking this up in his absence.

* update language based on GDPR Commit Group input

add inclusive language for all v1.x strings and reference transition period