Update publisher restrictions format
added "purpose 63" (to be discussed) in order to signal that this override should apply to all purposes. an "override all" is the more common scenario but is currently requiereing 10 restriction sections for each of the 10 purposes. an "all purposes" type can reduce the length of the tc string drastically while still being very explicit in the meaning.