clearly handle no cookie use
Currently, IAB Europe, the managing organization (MO) for TCF is using the value -100 as a proxy to indicate a GVL vendor does not use cookies of any kind (including session cookies). This pull request is meant to add a clear indication to CMPs and subsequently to users that a vendor does not use cookies of any kind.