Focusing the PR only on adding the FAQ
Initially, this PR was mainly focused on the examples (updates, adding new ones...) to cover more scenario in order to help the vendors. As there are many scenarios and to avoid creating misunderstanding with the previous ones, no major changes in the examples and the new FAQ is designed to be the place to addresses questions from vendors.